pastish_user built from shell params
hostname(1) is too finicky, the -f option pushes it
to query the dns (but it may omit the domain part
without it), which can hang/timeout the tests.

while i'm at it, $USER defaults to "anonymous" and
$HOSTNAME to "unknown".

to get some exposure, pastish_user disappears from
tests/*/.pastish; it was there to avoid having to
mock $USER and `hostname -f`, which (avoidance)
proved to be ultimately as futile as i expected.

tests/cmd introduced ITMT provided a convenient
hook to hang the cooked variables off of.